% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{PLE_oa_PSmatching-package}
\alias{PLE_oa_PSmatching}
\alias{PLE_oa_PSmatching-package}
\title{PLE_oa_PSmatching: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
